backend/requirements.txt
aiosqlite>=0.20.0
alembic>=1.14.0
anthropic>=0.40.0
apscheduler>=3.10.4
chromadb>=0.5.0
fastapi>=0.115.0
gitpython>=3.1.0
httpx>=0.27.0
pandas>=2.2.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.6.0
pydantic>=2.9.0
pypdf>=5.0.0
pytest-asyncio>=0.24.0
pytest>=8.3.0
python-docx>=1.1.0
python-dotenv>=1.0.0
python-jobspy>=1.1.80
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12
sqlalchemy>=2.0.36
trafilatura>=1.12.0
uvicorn[standard]>=0.32.0
